FoundrAI2_0/backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
huggingface_hub
reportlab
